# lab1/openapi-service-1.yml
openapi: '3.0.0'
info:
  title: "СОА спецификация OpenAPI"
  description: Routes сервис для работы с маршрутами (REST API)


    Беляев Михаил и Дениченко Александр


    Группа P3412


    Вариант 8801
  version: 1.0.0
servers:
  - url: http://localhost:27275/api/v1


tags:
  - name: "Routes"

paths:
  /routes:
    get:
      tags:
          - "Routes"
      summary: "Получить все маршруты с фильтрацией, сортировкой и пагинацией"
      operationId: "getRoutes"
      parameters:
        - name: sort
          in: query
          description: "Поля для сортировки. Формат: `fieldName_asc` или `fieldName_desc`. Несколько полей через запятую."
          required: false
          schema:
            type: string
            example: "name_asc,distance_desc"
        - name: filter
          in: query
          description: "Параметры для фильтрации."
          required: false
          style: form
          explode: true
          schema:
            type: object
            properties:
              name: 
                type: string
                description: "Фильтрация по названию"
                example: "Москва"
              distance:
                type: integer
                format: int64
                description: "Фильтрация по дистанции"
                example: 100
              "coordinates.x":
                type: number
                format: double
                description: "Фильтрация по координате X"
                example: 100.0
              "coordinates.y":
                type: integer
                format: int64
                description: "Фильтрация по координате Y"
                example: 100
              "from.x":
                type: number
                format: double
                description: "Фильтрация по координате X начальной точки"
                example: 100.0
              "from.y":
                type: number
                format: double
                description: "Фильтрация по координате Y начальной точки"
                example: 100.0
              "from.name":
                type: string
                description: "Фильтрация по названию начальной точки"
                example: "Москва"
              "to.x":
                type: number
                format: double
                description: "Фильтрация по координате X конечной точки"
                example: 100.0
              "to.y":
                type: number
                format: double
                description: "Фильтрация по координате Y конечной точки"
                example: 100.0
              "to.name":
                type: string
                description: "Фильтрация по названию конечной точки"
                example: "Москва"
              creationDate:
                type: string
                format: date-time
                description: "Фильтрация по дате создания"
                example: "2024-01-01T00:00:00Z"
              priority:
                type: string
                description: "Фильтрация по важности маршрута"
                enum: ["LOW", "MEDIUM", "HIGH"]
                example: "HIGH"
        - name: operation
          in: query
          description: "Операторы сравнения для полей фильтрации."
          required: false
          style: form
          explode: true
          schema:
            type: object
            properties:
              name_op: 
                type: string
                description: "Оператор для поля 'name'"
                enum: ["==", "!=", "~"]
                default: "=="
                example: "=="
              distance_op:
                type: string
                description: "Оператор для поля 'distance'"
                enum: ["==", "!=", ">", ">=", "<", "<="]
                default: "=="
                example: "=="
              "coordinates.x_op":
                type: string
                description: "Оператор для поля 'coordinates.x'"
                enum: ["==", "!=", ">", ">=", "<", "<="]
                default: "=="
                example: "=="
              "coordinates.y_op":
                type: string
                description: "Оператор для поля 'coordinates.y'"
                enum: ["==", "!=", ">", ">=", "<", "<="]
                default: "=="
                example: "=="
              "from.x_op":
                type: string
                description: "Оператор для поля 'from.x'"
                enum: ["==", "!=", ">", ">=", "<", "<="]
                default: "=="
                example: "=="
              "from.y_op":
                type: string
                description: "Оператор для поля 'from.y'"
                enum: ["==", "!=", ">", ">=", "<", "<="]
                default: "=="
                example: "=="
              "from.name_op":
                type: string
                description: "Оператор для поля 'from.name'"
                enum: ["==", "!=", "~"]
                default: "=="
                example: "=="
              "to.x_op":
                type: string
                description: "Оператор для поля 'to.x'"
                enum: ["==", "!=", ">", ">=", "<", "<="]
                default: "=="
                example: "=="
              "to.y_op":
                type: string
                description: "Оператор для поля 'to.y'"
                enum: ["==", "!=", ">", ">=", "<", "<="]
                default: "=="
                example: "=="
              "to.name_op":
                type: string
                description: "Оператор для поля 'to.name'"
                enum: ["==", "!=", "~"]
                default: "=="
                example: "=="
              creationDate_op:
                type: string
                description: "Оператор для поля 'creationDate'"
                enum: ["==", "!=", ">", ">=", "<", "<="]
                default: "=="
                example: "=="
              priority_op:
                type: string
                description: "Оператор для поля 'priority'"
                enum: ["==", "!=", ">", ">=", "<", "<="]
                default: "=="
                example: "=="
        - name: page
          in: query
          description: "Пагинация по номеру страницы (начиная с 1)"
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          description: "Количество элементов на странице"
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1
            maximum: 50
            default: 10
      responses:
          '200':
            description: "Успешный ответ с массивом маршрутов"
            headers:
              X-Total-Count:
                description: "Общее количество записей"
                schema:
                  type: integer
                  example: 150
            content:
              application/xml:
                schema:
                  type: array
                  xml:
                    name: RouteList
                    wrapped: true
                  items:
                    $ref: '#/components/schemas/Route'
          '400':
            description: "Некорректные данные в URL запросе"
            content:
              application/xml:
                schema:
                  $ref: '#/components/schemas/BadRequestError'
          '500':
            description: "Внутренняя ошибка сервера"
            content:
              application/xml:
                schema:
                  $ref: '#/components/schemas/InternalServerError'
          default:
            description: "Прочие сообщения об ошибках"
            content:
              application/xml:
                schema:
                  $ref: '#/components/schemas/Error'
    post:
      tags:
        - "Routes"
      summary: "Добавить новый маршрут"
      description: "Создает новый объект Route на основе переданных данных в теле запроса."
      operationId: "createRoute"
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              $ref: '#/components/schemas/RouteInput'
      responses:
        '201':
          description: "Маршрут успешно создан."
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Route'
        '400':
          description: "Некорректные данные в URL запросе"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/BadRequestError'
        '422':
          description: "Некорректные данные в теле запроса"
          content:
            application/xml:
              schema:
                  $ref: '#/components/schemas/UnprocessableEntityError'
        '500':
          description: "Внутренняя ошибка сервера"
          content:
            application/xml:
              schema:
                  $ref: '#/components/schemas/InternalServerError'
        default:
          description: "Прочие сообщения об ошибках"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Error'
    
  /routes/{id}:
    get:
      tags:
        - "Routes"
      summary: "Получить маршрут по ID"
      description: "Возвращает объект Route по указанному ID."
      operationId: "getRouteById"
      parameters:
        - name: id
          in: path
          description: "Идентификатор маршрута"
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
            exclusiveMinimum: true
            example: 1
      responses:
        '200':
          description: "Успешный ответ с объектом Route"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Route'
        '400':
          description: "Некорректные данные в URL запросе"
          content:
            application/xml:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BadRequestError'
                  - type: object
                    properties:
                      message:
                        example: "Некорректные данные в URL запросе. Тип поля id должен быть integer."
        '404':
          description: "Маршрут с указанным ID не найден"
          content:
            application/xml:
              schema:
                  $ref: '#/components/schemas/NotFoundError'
        '500':
          description: "Внутренняя ошибка сервера"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        default:
          description: "Прочие сообщения об ошибках"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - "Routes"
      summary: "Обновить существующий маршрут"
      description: "Обновляет объект Route."
      operationId: "updateRoute"
      requestBody:
        description: "Обновленные данные для маршрута."
        required: true
        content:
          application/xml:
            schema:
              $ref: '#/components/schemas/RouteInput'
      parameters:
        - name: id
          in: path
          description: "Идентификатор маршрута"
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
            exclusiveMinimum: true
            example: 1
      responses:
        '200':
          description: "Успешный ответ с объектом Route"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Route'
        '400':
          description: "Некорректные данные в URL запросе"
          content:
            application/xml:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BadRequestError'
                  - type: object
                    properties:
                      message:
                        example: "Некорректные данные в URL запросе. Тип поля id должен быть integer."
        '404':
          description: "Маршрут с указанным ID не найден"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '422':
          description: "Некорректные данные в теле запроса"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/UnprocessableEntityError'
        '500':
          description: "Внутренняя ошибка сервера"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        default:
          description: "Прочие сообщения об ошибках"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - "Routes"
      summary: "Удалить маршрут"
      description: "Удаляет объект Route."
      operationId: "deleteRoute"
      parameters:
        - name: id
          in: path
          description: "Идентификатор маршрута"
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
            exclusiveMinimum: true
            example: 1
      responses:
        '204':
          description: "Маршрут успешно удален"
        '400':
          description: "Некорректные данные в URL запросе"
          content:
            application/xml:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BadRequestError'
                  - type: object
                    properties:
                      message:
                        example: "Некорректные данные в URL запросе. Тип поля id должен быть integer."
        '404':
          description: "Маршрут не найден"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: "Внутренняя ошибка сервера"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        default:
          description: "Прочие сообщения об ошибках"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Error'

  /routes/max/from:
    get:
      tags:
        - "Routes"
      summary: "Получить один маршрут с максимальным значением поля 'from'"
      description: "Возвращает один (любой) объект, значение поля `from` которого является максимальным. Маршрут считается максимальным, если его значение `max(x, y)` больше, чем у остальных маршрутов. Если таких маршрутов несколько, возвращается первый найденный (с наименьшим `ID`)."
      operationId: "getMaxByFrom"
      responses:
        '200':
          description: "Успешный ответ с объектом Route"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Route'
        '400':
          description: "Некорректные данные в URL запросе"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/BadRequestError'
        '404':
          description: "В коллекции нет ни одного маршрута"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: "Внутренняя ошибка сервера"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        default:
          description: "Прочие сообщения об ошибках"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Error'

  /routes/distances/unique:
    get:
      tags:
        - "Routes"
      summary: "Получить массив уникальных значений поля distance"
      operationId: "getUniqueDistances"
      responses:
        '200':
          description: "Массив уникальных дистанций"
          content:
            application/xml:
              schema:
                type: array
                xml:
                  name: DistanceList
                  wrapped: true
                uniqueItems: true
                items:
                  type: integer
                  format: int64
                  description: "Уникальное значение дистанции"
                  xml:
                    name: Distance
                example: [100, 250, 500, 1000]
        '400':
          description: "Некорректные данные в URL запросе"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/BadRequestError'
        '500':
          description: "Внутренняя ошибка сервера"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        default:
          description: "Прочие сообщения об ошибках"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Error'

 
components:
  schemas:
    
    Coordinates:
      type: object
      description: "Координаты"
      xml:
        name: "Coordinates"
      properties:
        x:
          type: number
          format: double
          maximum: 716
          description: "X координата. Максимальное значение поля: 716"
          example: 100.0
        y:
          type: integer
          format: int64
          description: "Y координата. Поле не может быть null"
          example: 100
      required:
        - y

    Location:
      type: object
      description: "Местоположение с координатами и его названием"
      xml:
        name: "Location"
      properties:
        x:
          type: number
          format: double
          description: "X координата. Не может быть null."
          example: 100.0
        y:
          type: number
          format: double
          description: "Y координата."
          example: 100.0
        name:
          type: string
          minLength: 1
          description: "Название локации. Не может быть пустой, не может быть null."
          example: "Москва"
      required:
        - x
        - name
    
    Priority:
      type: string
      description: "Важность маршрута"
      enum:
        - "LOW"
        - "MEDIUM"
        - "HIGH"
      example: "HIGH"
      xml:
        name: "priority"

    RouteInput:
      type: object
      description: "Описывает маршрут с начальной и конечной точками."
      xml:
        name: "Route"
      properties:
        name:
          type: string
          description: "Название маршрута"
          minLength: 1
          example: "Москва"
        from:
          allOf:
            - $ref: '#/components/schemas/Location'
            - type: object
              xml:
                name: from
        to:
          allOf:
            - $ref: '#/components/schemas/Location'
            - type: object
              xml:
                name: to
        coordinates:
          allOf:
            - $ref: '#/components/schemas/Coordinates'
            - type: object
              xml:
                name: coordinates
        priority:
          $ref: '#/components/schemas/Priority'
        distance:
          type: integer
          format: int64
          description: "Длина маршрута. Должна быть больше 1."
          minimum: 1
          exclusiveMinimum: true
          nullable: true
          example: 100
      required:
        - name
        - coordinates
        - from
        - to

    Route:
      allOf:
        - type: object
          xml:
            name: "Route"
          properties:
            id:
              type: integer
              format: int64
              description: "Уникальный идентификатор (автоматически генерируется). Значение поля должно быть больше 0."
              readOnly: true
              minimum: 0
              exclusiveMinimum: true
              example: 1
            creationDate:
              type: string
              format: date-time
              description: "Дата и время создания (автоматически генерируется)."
              readOnly: true
              example: "2025-01-01T00:00:00Z"
        - $ref: '#/components/schemas/RouteInput'

    Error:
      type: object
      description: "Ошибка"
      xml:
        name: "Error"
      properties:
        status:
          type: integer
          format: int64
          description: "Статус ошибки"
        message:
          type: string
          description: "Сообщение об ошибке"
      required:
        - status
        - message
    
    InternalServerError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            status:
              example: 500
            message:
              example: "Внутренняя ошибка сервера"

    NotFoundError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            status:
              example: 404
            message:
              example: "Маршрут с указанным ID не найден"
    
    BadRequestError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            status:
              example: 400
            message:
              example: "Некорректные данные в URL запросе. Тип поля distance должен быть integer."
    
    UnprocessableEntityError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            status:
              example: 422
            message:
              example: "Некорректные данные в теле запроса. Тип поля distance должен быть integer."